code/BACKEND-API-GATEWAY/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Consul KV integration
python-consul2==0.1.5

# HTTP client for service communication
httpx==0.25.0

# Request handling
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1